backend/requirements_compact.txt
numpy==2.1.3
requests==2.32.3
beautifulsoup4==4.13.3
urllib3==2.3.0
python-dotenv==1.1.0
websockets==15.0.1
sentence-transformers==4.0.2
faiss-cpu==1.10.0
torch==2.6.0
PyPDF2==3.0.1
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.51
langchain-text-splitters==0.3.8
groq==0.22.0
docling==2.28.4
docling-core==2.25.0
docling-ibm-models==3.4.1
docling-parse==4.0.0
